EvaporationModel documentation

EvaporationModel documentation

Evaporation

Soil

Utils

Evaporation

EvaporationModel.compute_g_from_r_net — Method

compute_g_from_r_net(r_net, lai)

Compute the ground heat flux [W/m²] from net radiation (and LAI)

Arguments

r_net: The net radiation [W/m²].

lai: The leaf area index [m²/m²].

Details

This implementation follows the approach of the METRIC model. See Equation 27 of Allen et al., 2007

EvaporationModel.penman_monteith — Method

penman_monteith(t_air, p_surf, r_net, vpd, r_a, r_s; g = 0.0, kwargs...)

Compute evaporation (ET) and latent heat flux (LE) using the Penman-Monteith equation.

Arguments

t_air: Air temperature [K].

p_surf: Surface pressure [Pa].

r_net: Net radiation [W/m²].

vpd: Vapor pressure deficit [Pa].

r_a: Aerodynamic resistance [s/m].

r_s: Surface resistance [s/m].

g: Soil heat flux [W/m²]. Default is 0.0.

kwargs: Additional keyword arguments.

Returns

et: Potential evapotranspiration [kg/(m^2 * s)].

le: Latent heat flux [W/m²].

Soil

EvaporationModel.compute_a — Method

compute_a(perc_clay)

Compute a, a parameter for for w_geq calculation, of clay in the soil. See equation 35 of Noilhan & Mahfouf, 1996.

Arguments

perc_clay: The percentage of clay in the soil [%].

EvaporationModel.compute_b — Method

compute_b(::Val{:clay}, perc_clay)
compute_b(::Val{:van_genuchten}, n)

Compute b, the Brooks-Corey/Clapp-Hornberger parameter (see equation 1 of Clapp & Hornberger for its definition), based on percentage clay or the van Genuchten paramter n.

Arguments

approach: Either Val(:clay) or Val(:van_genuchten).

perc_clay: The percentage of clay in the soil [%], only for approach = Val(:clay).

n: The Van Genuchten parameter n, only for approach = Val(:van_genuchten).

Details

For the Val(:clay) approach, Equation (30) of Noilhan & Mahfouf, 1996 is used.

For the Val(:van_genuchten) approach, the parameter equivalence between the Brooks-Corey and van Genuchten, is based on Morel-Seytoux et al., 1996. Note that in this paper, M is equivalent to b.

EvaporationModel.compute_c_1 — Method

compute_c_1(w_g, w_sat, b, c_1_sat)

Compute force coefficient c_1 of force restore framework for soil mositure. See equation 20 of Noilhan & Mahfouf, 1996.

Arguments

w_g: Surface soil moisutre [m³ m⁻³]

w_sat: Saturated soil moisture [m³ m⁻³]

b: the Brooks-Corey/Clapp-Hornberger parameter, see compute_b

c_1_sat: See computec1_sat

EvaporationModel.compute_c_1_sat — Method

compute_c_1_sat(perc_clay)

Compute the value for force-restore coefficient c_1 when w_g = w_sat, c_2_ref based on the percentage of clay in the soil. See equation 32 of Noilhan & Mahfouf, 1996.

Arguments

perc_clay: The percentage of clay in the soil.

Returns

c_1_sat: The computed value of c_1_sat.

EvaporationModel.compute_c_2 — Method

compute_c_2(w_2, w_sat, c2_ref)

Compute restore coefficient c_2 of force restore framework for soil moisture See equation 21 of Noilhan & Mahfouf, 1996.

Arguments

w_2: The second layer soil mositure [m³ m⁻³]

w_sat: The saturated soil moisture [m³ m⁻³]

c2_ref: See computec2_ref

EvaporationModel.compute_c_2_ref — Method

compute_c_2_ref(perc_clay)

Compute the value for force-restore coefficient c_2 when w_2 = 0.5 w_sat, c_2_ref based on the percentage of clay in the soil. See equation 33 of Noilhan & Mahfouf, 1996.

Arguments

perc_clay: The percentage of clay in the soil [%].

EvaporationModel.compute_c_3 — Method

compute_c_3(perc_clay)

Compute the coefficient for graviational drainage c_3 based on the percentage of clay in the soil. See equation 34 of Noilhan & Mahfouf, 1996.

Arguments

perc_clay: The percentage of clay in the soil [%].

EvaporationModel.compute_p — Method

compute_p(perc_clay)

Compute p, a parameter for for w_geq calculation, based on the percentage of clay in the soil. See equation 36 of Noilhan & Mahfouf, 1996.

Arguments

perc_clay: The percentage of clay in the soil [%].

EvaporationModel.compute_w_geq — Method

compute_w_geq(w_2, w_sat, a, p)

Compute w_geq, the equilibrium surface soil moisture (i.e. when capillary and gravitational forces are in equilibrium). See equation 19 of Noilhan & Mahfouf, 1996.

Arguments

w_2: The second layer soil moisture [m³ m⁻³].

w_sat: The saturated soil moisture [m³ m⁻³].

a: Clapp-Hornberger parameter a (see compute_a)

p: Clapp-Hornberger parameter a (see compute_p)

Utils

EvaporationModel.fourier_series — Method

fourier_series(t, coeffs, ω)

Function that gives the Fourier series, as presented in equation 5 of Murray & Verhoef

$f(t) = a_0 + \sum_{n=1}^M \left( a_n \cos(n \omega t) + b_n \sin(n \omega t) \right)$

Arguments

t: Timestamp in seconds

coeffs::ComponentArray: array holding the coefficients

a0: Mean component

an: Cosine coefficients

bn: Sine coefficients

ω: Radial frequency [rad/s]