EvaporationModel documentation

EvaporationModel documentation

Config

Evaporation

Soil

Resistances

Utils

Config

EvaporationModel.VegetationParameters — Type

VegetationParameters(;...)

Default vegetation parameters for the Jarvis-Stewart model. The fields are:

vegtype: The type of vegetation (e.g., Crops, ShortGrass, etc.)

r_smin: Minimum stomatal resistance [s/m]

g_d: Coefficient relating vapour pressure deficit to stomatal resistance [Pa⁻¹]

Based on get_default_value function, values of r_smin and g_d are set based on vegtype. Default values can be overridden by passing them as keyword arguments.

Examples

using EvaporationModel
# Get default parameters
params_default = VegetationParameters(vegtype=Crops())
# Adapt a default value
params = VegetationParameters(vegtype=Crops(), r_smin=300.0) 
params.r_smin == 300.0

# output

true

EvaporationModel.VegetationType — Type

VegetationType

Abstract type defined to represent different types of vegetation.

Below example provided on how to check all the subtypes (which are structs):

subtypes(VegetationType)

EvaporationModel.get_default_value — Method

get_default_value(vegtype::VegetationType)

Assign default values for minimal stomatal resistance (r_smin) and coefficient relating vapour pressure deficit to stomatal resistance (g_d) based on vegetation type.

The values come from Table 8.1 of the IFS Cy49r1 documentation Part IV: Phyiscal processes. Not part of public API, only used in VegetationParameters.

Evaporation

EvaporationModel.penman_monteith — Method

penman_monteith(t_air, p_surf, r_net, vpd, r_a, r_s; g = 0.0, kwargs...)

Compute evaporation (ET) and latent heat flux (LE) using the Penman-Monteith equation.

Arguments

t_air: Air temperature [K].

p_surf: Surface pressure [Pa].

r_net: Net radiation [W/m²].

vpd: Vapor pressure deficit [Pa].

r_a: Aerodynamic resistance [s/m].

r_s: Surface resistance [s/m].

g: Soil heat flux [W/m²]. Default is 0.0.

kwargs: Additional keyword arguments.

Returns

et: Potential evapotranspiration [kg/(m^2 * s)].

le: Latent heat flux [W/m²].

Soil

EvaporationModel.c_1 — Method

c_1(w_1, w_sat, b, c_1sat)

Compute force coefficient c_1 of force restore framework for soil mositure. See equation 20 of Noilhan & Mahfouf, 1996.

Arguments

w_1: Surface soil moisutre [m³ m⁻³]

w_sat: Saturated soil moisture [m³ m⁻³]

b: the Brooks-Corey/Clapp-Hornberger parameter, see compute_b

c_1sat: See c_1sat

EvaporationModel.c_1sat — Method

c_1sat(x_clay)

Compute the value for force-restore coefficient c_1 when w_g = w_sat. See equation 32 of Noilhan & Mahfouf, 1996.

Arguments

x_clay: The percentage of clay in the soil.

Returns

c_1sat: The computed value of c_1sat.

EvaporationModel.c_2 — Method

c_2(w_2, w_sat, c2_ref)

Compute restore coefficient c_2 of force restore framework for soil moisture See equation 21 of Noilhan & Mahfouf, 1996.

Arguments

w_2: The second layer soil mositure [m³ m⁻³]

w_sat: The saturated soil moisture [m³ m⁻³]

c_2ref: See c_2ref

EvaporationModel.c_2ref — Method

c_2ref(x_clay)

Compute the value for force-restore coefficient c_2 when w_2 = 0.5 w_sat, c_2ref based on the percentage of clay in the soil. See equation 33 of Noilhan & Mahfouf, 1996.

Arguments

x_clay: The percentage of clay in the soil [%].

EvaporationModel.c_3 — Method

c_3(x_clay)

Compute the coefficient for graviational drainage c_3 based on the percentage of clay in the soil. See equation 34 of Noilhan & Mahfouf, 1996.

Arguments

x_clay: The percentage of clay in the soil [%].

EvaporationModel.compute_a — Method

compute_a(x_clay)

Compute a, a parameter for for w_geq calculation, of clay in the soil. See equation 35 of Noilhan & Mahfouf, 1996.

Arguments

x_clay: The percentage of clay in the soil [%].

EvaporationModel.compute_b — Method

compute_b(::Val{:clay}, x_clay)
compute_b(::Val{:van_genuchten}, n)

Compute b, the Brooks-Corey/Clapp-Hornberger parameter (see equation 1 of Clapp & Hornberger for its definition), based on percentage clay or the van Genuchten paramter n.

Arguments

approach: Either Val(:clay) or Val(:van_genuchten).

x_clay: The percentage of clay in the soil [%], only for approach = Val(:clay).

n: The Van Genuchten parameter n, only for approach = Val(:van_genuchten).

Details

For the Val(:clay) approach, Equation (30) of Noilhan & Mahfouf, 1996 is used.

For the Val(:van_genuchten) approach, the parameter equivalence between the Brooks-Corey and van Genuchten, is based on Morel-Seytoux et al., 1996. Note that in this paper, M is equivalent to b.

EvaporationModel.compute_p — Method

compute_p(x_clay)

Compute p, a parameter for for w_geq calculation, based on the percentage of clay in the soil. See equation 36 of Noilhan & Mahfouf, 1996.

Arguments

x_clay: The percentage of clay in the soil [%].

EvaporationModel.w_geq — Method

w_geq(w_2, w_sat, a, p)

Compute w_geq, the equilibrium surface soil moisture (i.e. when capillary and gravitational forces are in equilibrium). See equation 19 of Noilhan & Mahfouf, 1996.

Arguments

w_2: The second layer soil moisture [m³ m⁻³].

w_sat: The saturated soil moisture [m³ m⁻³].

a: Clapp-Hornberger parameter a (see compute_a)

p: Clapp-Hornberger parameter a (see compute_p)

Resistances

EvaporationModel.soil_aerodynamic_resistance — Method

soil_aerodynamic_resistance(::Choudhury1988soil, ustar, η, h, d_c, z_0mc, z_0ms)

Calculate soil aerodynamic resistance, which is between the soil surface and the canopy source height ($z_m = z_{0mc} + d_{c}$)

Arguments

approach: The approach to use for calculating soil aerodynamic resistance

ustar: Friction velocity [m/s]

η: Extinction coefficient of $K$ in canopy [-]

h: Canopy height [m]

d_c: Canopy displacement height [m]

z_0mc: Roughness length for momentum transfer for canopy [m]

z_0ms: Roughness length for momentum transfer for soil [m]

Returns

r_as: Soil aerodynamic resistance [s/m]

Details

With approach = Choudhury1988soil(), equation (25) of Choudhury and Monteith (1988) is used.

EvaporationModel.surface_resistance — Method

surface_resistance(::JarvisStewart, s_in, vpd, t_air, w_2, w_fc, w_wilt, lai, g_d, r_smin)

Calculate surface resistance

Arguments

approach: The approach to use for calculating surface resistance

s_in: Incoming solar radiation [W/m²]

vpd: Vapor pressure deficit [Pa]

t_air: Air temperature [K]

w_2: Root zone soil moisture [m³/m³]

w_fc: Soil moisture at field capacity [m³/m³]

w_wilt: Soil moisture at wilting point [m³/m³]

lai: Leaf area index [m²/m²]

g_d: Parameter relating vpd to stomatal conductance [Pa⁻¹]

r_smin: Minimum stomatal resistance [s/m]

Returns

r_s: Surface resistance [s/m]

Details

With approach = JarvisStewart(), the Jarvis-Stewart model is used as given in equations (8.9), (8.10), (8.11) and (8.14) of the IFS Cy47r1 documentation Part IV: Physical processes. The temperature constraint (f_4) comes from Noilhan and Planton (1989) equation (37), with the default value of $T_{opt}$ set to 298.0 K

Utils

EvaporationModel.compute_amplitude_and_phase — Method

compute_amplitude_and_phase(an::AbstractVector, bn::AbstractVector)

Translate the fourier coefficients from sin-cos form (see fourier_series) to amplitude phase form:

$f(t) = a_0 + \sum_{n=1}^M \left( a_{bn} \sin(n \omega t + \phi) \right)$

Specifically, it returns both a_{bn} and \phi vectors. FYI: proof of conversion found here. Crucial to use atan2 function

EvaporationModel.fourier_series — Method

fourier_series(t, coeffs, ω)

Function that gives the Fourier series, as presented in equation 5 of Murray & Verhoef

$f(t) = a_0 + \sum_{n=1}^M \left( a_n \cos(n \omega t) + b_n \sin(n \omega t) \right)$

Arguments

t: Timestamp in seconds

coeffs::ComponentArray: array holding the coefficients

a0: Mean component

an: Cosine coefficients

bn: Sine coefficients

ω: Radial frequency [rad/s]